requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
langchain
langchain_classic
langchain-community
langchain-openai
chromadb
pymupdf
langdetect
tiktoken
twilio
langchain-ollama
langchain-huggingface
sentence-transformers
langchain_chroma
langchain_core
